Asp.Net_Core_Personal_Website
There are 3 basic panels on this personal website developed with .Net Core.
While the default panel includes information such as your education, experiences, references, and projects, a panel has been created in the author panel for people who will register to your system and communicate through this system. In this panel, people will be able to access transactions such as messaging, announcements, editing profile information, and changing passwords.
The admin panel is the main part of the project.